linux-sandbox-rs: give includes priority over excludes
The motivation for this was to allow more specific include paths in
targets to take priority over excludes from targets (i.e. a perl
filegroup that wishes to include a folder named `CORE` but to exclude
`libperl.so` and another target that includes just `libperl.so` —
depending on both should map `libperl.so` into the sandbox, overriding
the original exclude).

However, there's another impedient to this:

Bazel seems to silently elide artifacts that are subsets of other
artifacts; i.e. since the include on the first filegroup (`CORE/`) is a
superset of the include on the `libperl.so` filegroup
(`CORE/libperl.so`) when a target depends on the `CORE` filegroup, it's
as if the dep on the latter fileroup is dropped. And so the include from
the second fileroup (i.e. a bind mount to `libperl.so` that we were
hoping would override our exclude here...) doesn't make it in.

TODO: it'd be nice to have bazel not do this elision when there are
extra excludes/includes at play...